libs/shared/iam/data-access/src/lib/+state/iam.facade.spec.ts
import { NgModule } from '@angular/core';
import { TestBed } from '@angular/core/testing';
import { cold, readFirst } from '@nrwl/angular/testing';
import { RouterTestingModule } from '@angular/router/testing';

import { EffectsModule } from '@ngrx/effects';
import { StoreModule, Store } from '@ngrx/store';

import { NxModule } from '@nrwl/angular';

import { IAM } from './iam.models';
import { IamEffects } from './iam.effects';
import { IamFacade } from './iam.facade';

import * as IamSelectors from './iam.selectors';
import * as IamActions from './iam.actions';
import { IAM_FEATURE_KEY, State, initialState, reducer } from './iam.reducer';
import { IamApiService } from '../api/iam-api.service';

interface TestSchema {
  iam: State;
}

describe('IamFacade', () => {
  let facade: IamFacade;
  let store: Store<TestSchema>;
  let iamApiService: IamApiService;

  const iamApiServiceMock = jest.fn<Partial<IamApiService>, []>(() => ({
    signup: jest.fn(),
  }));
  const createUserProfileEntity = (
    id: string,
    firstName = '',
    lastName = '',
    email = '',
    isActive = true
  ) =>
    ({
      id,
      firstName: firstName || `firstName-${id}`,
      lastName: lastName || `lastName-${id}`,
      email: email || `${firstName}.${lastName}-@email.com`,
      isActive: isActive,
    } as IAM.UserProfile);

  beforeEach(() => {});

  describe('used in NgModule', () => {
    beforeEach(() => {
      @NgModule({
        imports: [
          StoreModule.forFeature(IAM_FEATURE_KEY, reducer),
          EffectsModule.forFeature([IamEffects]),
        ],
        providers: [IamFacade],
      })
      class CustomFeatureModule {}

      @NgModule({
        imports: [
          RouterTestingModule,
          NxModule.forRoot(),
          StoreModule.forRoot({}),
          EffectsModule.forRoot([]),
          CustomFeatureModule,
        ],
        providers: [{ provide: IamApiService, useValue: iamApiServiceMock() }],
      })
      class RootModule {}
      TestBed.configureTestingModule({ imports: [RootModule] });

      store = TestBed.inject(Store);
      facade = TestBed.inject(IamFacade);
      iamApiService = TestBed.inject(IamApiService);
    });

    /**
     * The initially generated facade::loadAll() returns empty array
     */
    it('signupUser() should return a selected item list with loaded == true', async (done) => {
      jest.spyOn(iamApiService, 'signup').mockReturnValue(cold('a|'));
      const { firstName, lastName, email } = createUserProfileEntity(
        'USER-AAA'
      );
      const signupPayload: IAM.UserSignUp = {
        firstName,
        lastName,
        email,
        password: 'Sup3RS3cr3TP4ssw0rd',
      };

      try {
        let userProfile = await readFirst(facade.userProfile$);
        let isLoaded = await readFirst(facade.loaded$);

        expect(userProfile).toBe(undefined);
        expect(isLoaded).toBe(false);

        facade.signupUser(signupPayload);

        userProfile = await readFirst(facade.userProfile$);
        isLoaded = await readFirst(facade.loaded$);

        // TODO: fix this test, there seems to be an issue with timing
        // expect(userProfile).toBeDefined();
        // expect(isLoaded).toBe(true);

        done();
      } catch (err) {
        done.fail(err);
      }
    });

    /*     it('loadAll() should return empty list with loaded == true', async (done) => {
      try {
        let list = await readFirst(facade.allIam$);
        let isLoaded = await readFirst(facade.loaded$);

        expect(list.length).toBe(0);
        expect(isLoaded).toBe(false);

        facade.dispatch(IamActions.loadIam());

        list = await readFirst(facade.allIam$);
        isLoaded = await readFirst(facade.loaded$);

        expect(list.length).toBe(0);
        expect(isLoaded).toBe(true);

        done();
      } catch (err) {
        done.fail(err);
      }
    });
 */
    /**
     * Use `loadIamSuccess` to manually update list
     */
    /*     it('allIam$ should return the loaded list; and loaded flag == true', async (done) => {
      try {
        let list = await readFirst(facade.allIam$);
        let isLoaded = await readFirst(facade.loaded$);

        expect(list.length).toBe(0);
        expect(isLoaded).toBe(false);

        facade.dispatch(
          IamActions.loadIamSuccess({
            iam: [createUserProfileEntity('AAA'), createUserProfileEntity('BBB')],
          })
        );

        list = await readFirst(facade.allIam$);
        isLoaded = await readFirst(facade.loaded$);

        expect(list.length).toBe(2);
        expect(isLoaded).toBe(true);

        done();
      } catch (err) {
        done.fail(err);
      }
    });
 */
  });
});
